Initial release
Working design, verified in a Tektronix TDS3054 oscilloscope. Most of the settings can be configured through the menus but the power-on hours cannot to my knowledge. These are stored in locations 7E0-7E4 of the DS1742W in the format of minutes powered up, for example 00 02 00 94 00 will result in a reading of 2187 hrs. The DS1744WP is the PowerCap version incorporating a replaceable snap on module containing the battery and crystal instead of the potted block used in the through-hole versions.